Fix: model policies on none resource actions
